+++ sdk/js/src/token_bridge/getOriginalAsset.ts
@@ -10,7 +10,7 @@ import { getIsWrappedAssetEth } from "./getIsWrappedAsset";
 
 export interface WormholeWrappedInfo {
   isWrapped: boolean;
-  chainId: ChainId;
+  chainId: number;
   assetAddress: Uint8Array;
 }
 
@@ -37,7 +37,7 @@ export async function getOriginalAssetEth(
       wrappedAddress,
       provider
     );
-    const chainId = (await token.chainId()) as ChainId; // origin chain
+    const chainId = await token.chainId(); // origin chain
     const assetAddress = await token.nativeContract(); // origin address
     return {
       isWrapped: true,
@@ -66,7 +66,7 @@ export async function getOriginalAssetTerra(
   try {
     const result: {
       asset_address: string;
-      asset_chain: ChainId;
+      asset_chain: number;
       bridge: string;
     } = await client.wasm.contractQuery(wrappedAddress, {
       wrapped_asset_info: {},