+++ explorer/src/components/ChainIcon.tsx
@@ -7,6 +7,7 @@ import polygonIcon from "../images/polygon.svg";
 import avalancheIcon from "../images/avalanche.svg";
 import oasisIcon from "../images/oasis.svg";
 import fantomIcon from "../images/fantom.svg";
+import auroraIcon from "../images/aurora.svg";
 import {
   ChainId,
   CHAIN_ID_AVAX,
@@ -17,6 +18,7 @@ import {
   CHAIN_ID_SOLANA,
   CHAIN_ID_TERRA,
   CHAIN_ID_FANTOM,
+  CHAIN_ID_AURORA,
 } from "@certusone/wormhole-sdk";
 import { chainEnums } from "../utils/consts";
 import { Box } from "@mui/material";
@@ -30,6 +32,7 @@ const chainIdToSrc = {
   [CHAIN_ID_AVAX]: avalancheIcon,
   [CHAIN_ID_OASIS]: oasisIcon,
   [CHAIN_ID_FANTOM]: fantomIcon,
+  [CHAIN_ID_AURORA]: auroraIcon,
 };
 
 const ChainIcon = ({ chainId }: { chainId: ChainId }) =>