+++ node/hack/repair_eth/repair_eth.go
@@ -6,8 +6,10 @@ import (
 	"encoding/json"
 	"flag"
 	"fmt"
+	"io"
 	"log"
 	"net/http"
+	"net/http/cookiejar"
 	"strconv"
 	"strings"
 	"time"
@@ -32,6 +34,7 @@ var etherscanAPIMap = map[vaa.ChainID]string{
 	vaa.ChainIDPolygon:   "https://api.polygonscan.com/api",
 	vaa.ChainIDOasis:     "https://explorer.emerald.oasis.dev/api",
 	vaa.ChainIDAurora:    "https://explorer.mainnet.aurora.dev/api",
+	vaa.ChainIDFantom:    "https://api.ftmscan.com/api",
 }
 
 var coreContractMap = map[vaa.ChainID]string{
@@ -41,6 +44,7 @@ var coreContractMap = map[vaa.ChainID]string{
 	vaa.ChainIDPolygon:   "0x7A4B5a56256163F07b2C80A7cA55aBE66c4ec4d7",
 	vaa.ChainIDOasis:     "0xfe8cd454b4a1ca468b57d79c0cc77ef5b6f64585", // <- converted to all lower case for easy compares
 	vaa.ChainIDAurora:    "0xa321448d90d4e5b0a732867c18ea198e75cac48e",
+	vaa.ChainIDFantom:    strings.ToLower("0x126783A6Cb203a3E35344528B26ca3a0489a1485"),
 }
 
 var (
@@ -49,6 +53,7 @@ var (
 	chain        = flag.String("chain", "ethereum", "Eth Chain name")
 	dryRun       = flag.Bool("dryRun", true, "Dry run")
 	step         = flag.Uint64("step", 10000, "Step")
+	showError    = flag.Bool("showError", false, "On http error, show the response body")
 )
 
 var (
@@ -101,7 +106,7 @@ type logResponse struct {
 	Result json.RawMessage `json:"result"`
 }
 
-func getCurrentHeight(chainId vaa.ChainID, ctx context.Context, c *http.Client, api, key string) (uint64, error) {
+func getCurrentHeight(chainId vaa.ChainID, ctx context.Context, c *http.Client, api, key string, showErr bool) (uint64, error) {
 	var req *http.Request
 	var err error
 	if chainId == vaa.ChainIDOasis || chainId == vaa.ChainIDAurora {
@@ -124,15 +129,23 @@ func getCurrentHeight(chainId vaa.ChainID, ctx context.Context, c *http.Client, 
 	var r struct {
 		Result string `json:"result"`
 	}
+	body, err := io.ReadAll(resp.Body)
+	if err != nil {
+		return 0, fmt.Errorf("failed to read response body: %w", err)
+	}
 
-	if err := json.NewDecoder(resp.Body).Decode(&r); err != nil {
+	if resp.StatusCode != http.StatusOK && showErr {
+		fmt.Println(string(body))
+	}
+
+	if err := json.Unmarshal(body, &r); err != nil {
 		return 0, fmt.Errorf("failed to decode response: %w", err)
 	}
 
 	return hexutil.DecodeUint64(r.Result)
 }
 
-func getLogs(chainId vaa.ChainID, ctx context.Context, c *http.Client, api, key, contract, topic0 string, from, to string) ([]*logEntry, error) {
+func getLogs(chainId vaa.ChainID, ctx context.Context, c *http.Client, api, key, contract, topic0 string, from, to string, showErr bool) ([]*logEntry, error) {
 	var req *http.Request
 	var err error
 	if chainId == vaa.ChainIDOasis || chainId == vaa.ChainIDAurora {
@@ -158,7 +171,16 @@ func getLogs(chainId vaa.ChainID, ctx context.Context, c *http.Client, api, key,
 
 	var r logResponse
 
-	if err := json.NewDecoder(resp.Body).Decode(&r); err != nil {
+	body, err := io.ReadAll(resp.Body)
+	if err != nil {
+		return nil, fmt.Errorf("failed to read response body: %w", err)
+	}
+
+	if resp.StatusCode != http.StatusOK && showErr {
+		fmt.Println(string(body))
+	}
+
+	if err := json.Unmarshal(body, &r); err != nil {
 		return nil, fmt.Errorf("failed to decode response: %w", err)
 	}
 
@@ -215,7 +237,14 @@ func main() {
 
 	ctx := context.Background()
 
-	currentHeight, err := getCurrentHeight(chainID, ctx, http.DefaultClient, etherscanAPI, *etherscanKey)
+	jar, err := cookiejar.New(nil)
+	if err != nil {
+		log.Fatalf("Error creating http cookiejar: %v", err)
+	}
+	httpClient := &http.Client{
+		Jar: jar,
+	}
+	currentHeight, err := getCurrentHeight(chainID, ctx, httpClient, etherscanAPI, *etherscanKey, *showError)
 	if err != nil {
 		log.Fatalf("Failed to get current height: %v", err)
 	}
@@ -288,7 +317,10 @@ func main() {
 
 	limiter := rate.NewLimiter(rate.Every(1*time.Second), 1)
 
-	c := &http.Client{Timeout: 5 * time.Second}
+	c := &http.Client{
+		Jar:     jar,
+		Timeout: 5 * time.Second,
+	}
 
 	ethAbi, err := abi2.JSON(strings.NewReader(abi.AbiABI))
 	if err != nil {
@@ -315,7 +347,7 @@ func main() {
 
 		log.Printf("Requesting logs from block %s to %s", from, to)
 
-		logs, err := getLogs(chainID, ctx, c, etherscanAPI, *etherscanKey, coreContract, tokenLockupTopic.Hex(), from, to)
+		logs, err := getLogs(chainID, ctx, c, etherscanAPI, *etherscanKey, coreContract, tokenLockupTopic.Hex(), from, to, *showError)
 		if err != nil {
 			log.Fatalf("failed to get logs: %v", err)
 		}